+++ bin/mcserv.dart
@@ -9,8 +9,6 @@ import 'package:intl/intl_standalone.dart';
 import 'package:intl/locale.dart';
 import 'package:logging/logging.dart';
 import 'package:mcserv/intl/localizations.dart';
-import 'package:mcserv/utils/fs_util.dart';
-import 'package:mcserv/utils/localizations_util.dart';
 import 'package:mcserv/utils/utils.dart';
 import 'package:path/path.dart' as path;