+++ lib/distributions/distribution.dart
@@ -1,12 +1,12 @@
 import 'package:file/file.dart';
 import 'package:mcserv/distributions/pl3x/purpur_distribution.dart';
 import 'package:mcserv/distributions/plain/bungeecord_distribution.dart';
-import 'package:mcserv/distributions/plain/velocity_distributions.dart';
 import 'package:meta/meta.dart';
 
 import 'paper/paper_mc_distribution.dart';
 import 'paper/travertine_distribution.dart';
 import 'paper/waterfall_distribution.dart';
+import 'paper/velocity_distribution.dart';
 
 abstract class Distribution {
   static final List<Distribution> all = [