+++ lib/utils/recommendation_util.dart
@@ -1,4 +1,4 @@
-import 'package:mcserv/utils/localizations_util.dart';
+import 'localizations_util.dart';
 
 String recommend(String recommendation, bool recommend) => recommend
     ? recommendation + ' (${localizations.recommended})'